Move third order quad test to large tests
